requirements.txt
Flask
PyMuPDF
Pillow
google-generativeai
python-dotenv
